SWITCHYARD-232: changed quickstarts to follow convention
