+++ demos/orders/src/test/java/org/switchyard/quickstarts/demos/orders/WebServiceTest.java
@@ -1,0 +1,44 @@
+/* 
+ * JBoss, Home of Professional Open Source 
+ * Copyright 2010-2011 Red Hat Inc. and/or its affiliates and other contributors
+ * as indicated by the @authors tag. All rights reserved. 
+ * See the copyright.txt in the distribution for a 
+ * full listing of individual contributors.
+ *
+ * This copyrighted material is made available to anyone wishing to use, 
+ * modify, copy, or redistribute it subject to the terms and conditions 
+ * of the GNU Lesser General Public License, v. 2.1. 
+ * This program is distributed in the hope that it will be useful, but WITHOUT A 
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+ * A PARTICULAR PURPOSE.  See the GNU Lesser General Public License for more 
+ * details. You should have received a copy of the GNU Lesser General Public 
+ * License, v.2.1 along with this distribution; if not, write to the Free 
+ * Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  
+ * 02110-1301, USA.
+ */
+
+package org.switchyard.quickstarts.demos.orders;
+
+import org.junit.Test;
+import org.switchyard.component.bean.config.model.BeanSwitchYardScanner;
+import org.switchyard.test.SwitchYardTestCase;
+import org.switchyard.test.SwitchYardTestCaseConfig;
+import org.switchyard.test.mixins.CDIMixIn;
+import org.switchyard.test.mixins.HTTPMixIn;
+import org.switchyard.transform.config.model.TransformSwitchYardScanner;
+
+@SwitchYardTestCaseConfig(
+        config = SwitchYardTestCaseConfig.SWITCHYARD_XML,
+        scanners = {BeanSwitchYardScanner.class, TransformSwitchYardScanner.class},
+        mixins = {CDIMixIn.class, HTTPMixIn.class})
+public class WebServiceTest extends SwitchYardTestCase {
+
+    @Test
+    public void invokeOrderWebService() throws Exception {
+        // Use the HttpMixIn to invoke the SOAP binding endpoint with a SOAP input (from the test classpath)
+        // and compare the SOAP response to a SOAP response resource (from the test classpath)...
+        getMixIn(HTTPMixIn.class).
+                postResourceAndTestXML("http://localhost:18001/OrderService", "/xml/soap-request.xml", "/xml/soap-response.xml");
+    }
+}
+
