+++ demos/orders/src/test/resources/xml/order.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<orders:order xmlns:orders="urn:switchyard-quickstart:transform-smooks:1.0">
+<order xmlns="urn:switchyard-quickstart-demo:orders:1.0">
     <orderId>PO-19838-XYZ</orderId>
     <itemId>BUTTER</itemId>
     <quantity>200</quantity>
-</orders:order>
+</order>
