+++ demos/orders/src/test/resources/xml/orderAck.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<orders:submitOrderResponse xmlns:orders="urn:switchyard-quickstart-demo:orders:1.0">
+	<orderAck>
+		<orderId>PO-19838-XYZ</orderId>
+		<accepted>true</accepted>
+		<status>Order Accepted</status>
+	</orderAck>
+</orders:submitOrderResponse>
