+++ demos/orders/src/test/resources/xml/soap-request.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<soap:Envelope xmlns:soap="http://schemas.xmlsoap.org/soap/envelope/">
+	<soap:Body>
+		<orders:submitOrder xmlns:orders="urn:switchyard-quickstart-demo:orders:1.0">
+			<order>
+				<orderId>PO-19838-XYZ</orderId>
+				<itemId>BUTTER</itemId>
+				<quantity>200</quantity>
+			</order>
+		</orders:submitOrder>
+	</soap:Body>
+</soap:Envelope>
