+++ demos/orders/src/test/resources/xml/soap-response.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<SOAP-ENV:Envelope xmlns:SOAP-ENV="http://schemas.xmlsoap.org/soap/envelope/">
+    <SOAP-ENV:Body>
+        <orders:submitOrderResponse xmlns:orders="urn:switchyard-quickstart-demo:orders:1.0">
+            <orderAck>
+                <orderId>PO-19838-XYZ</orderId>
+                <accepted>true</accepted>
+                <status>Order Accepted</status>
+            </orderAck>
+        </orders:submitOrderResponse>
+    </SOAP-ENV:Body>
+</SOAP-ENV:Envelope>
