--- m1app/pom.xml
@@ -1,109 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- 
- * JBoss, Home of Professional Open Source 
- * Copyright 2011 Red Hat Inc. and/or its affiliates and other contributors
- * as indicated by the @author tags. All rights reserved. 
- * See the copyright.txt in the distribution for a 
- * full listing of individual contributors.
- *
- * This copyrighted material is made available to anyone wishing to use, 
- * modify, copy, or redistribute it subject to the terms and conditions 
- * of the GNU Lesser General Public License, v. 2.1. 
- * This program is distributed in the hope that it will be useful, but WITHOUT A 
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A 
- * PARTICULAR PURPOSE.  See the GNU Lesser General Public License for more details. 
- * You should have received a copy of the GNU Lesser General Public License, 
- * v.2.1 along with this distribution; if not, write to the Free Software 
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, 
- * MA  02110-1301, USA.
- */ -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.switchyard.quickstarts</groupId>
-    <artifactId>switchyard-quickstarts-parent</artifactId>
-    <version>0.1.0-SNAPSHOT</version>
-  </parent>
-
-  <groupId>org.switchyard.quickstarts</groupId>
-  <artifactId>m1app</artifactId>
-  <version>0.1.0-SNAPSHOT</version>
-  <packaging>jar</packaging>
-
-  <name>Quickstart : M1 Demo App</name>
-  <dependencies>
-    <dependency>
-      <groupId>org.switchyard</groupId>
-      <artifactId>switchyard-api</artifactId>
-      <version>0.1.0-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>org.switchyard</groupId>
-      <artifactId>switchyard-plugin</artifactId>
-      <version>0.1.0-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>org.switchyard.components</groupId>
-      <artifactId>switchyard-component-bean</artifactId>
-      <version>0.1.0-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>org.switchyard.components</groupId>
-      <artifactId>switchyard-component-soap</artifactId>
-      <version>0.1.0-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>org.switchyard</groupId>
-      <artifactId>switchyard-test</artifactId>
-      <version>0.1.0-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.8.1</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>commons-httpclient</groupId>
-      <artifactId>commons-httpclient</artifactId>
-      <version>3.1</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-  <build>
-     <plugins>
-      <plugin>
-          <groupId>org.apache.maven.plugins</groupId>
-          <artifactId>maven-compiler-plugin</artifactId>
-          <version>2.3.2</version>
-          <configuration>
-              <source>1.6</source>
-              <target>1.6</target>
-              <debug>true</debug>
-              <showWarnings>true</showWarnings>
-              <showDeprecation>true</showDeprecation>
-          </configuration>
-      </plugin>
-      <plugin>
-          <groupId>org.switchyard</groupId>
-          <artifactId>switchyard-plugin</artifactId>
-          <version>0.1.0-SNAPSHOT</version>
-          <configuration>
-            <scannerClassNames>
-                <param>org.switchyard.component.bean.config.model.BeanSwitchYardScanner</param>
-                <param>org.switchyard.transform.config.model.TransformSwitchYardScanner</param>
-            </scannerClassNames>
-          </configuration>
-          <executions>
-            <execution>
-                <goals>
-                    <goal>configure</goal>
-                </goals>
-            </execution>
-          </executions>
-      </plugin>
-     </plugins>
-  </build>
-</project>
