--- m1app/src/test/resources/xml/order.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<order xmlns="urn:switchyard-quickstarts:m1app:1.0">
-    <orderId>PO-19838-XYZ</orderId>
-    <itemId>BUTTER</itemId>
-    <quantity>200</quantity>
-</order>
