--- m1app/src/test/resources/xml/orderAck.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<m1:submitOrderResponse xmlns:m1="urn:switchyard-quickstarts:m1app:1.0">
-	<orderAck>
-		<orderId>PO-19838-XYZ</orderId>
-		<accepted>true</accepted>
-		<status>Order Accepted</status>
-	</orderAck>
-</m1:submitOrderResponse>
