--- m1app/src/test/resources/xml/soap-response.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<SOAP-ENV:Envelope xmlns:SOAP-ENV="http://schemas.xmlsoap.org/soap/envelope/">
-    <SOAP-ENV:Body>
-        <m1:submitOrderResponse xmlns:m1="urn:switchyard-quickstarts:m1app:1.0">
-            <orderAck>
-                <orderId>PO-19838-XYZ</orderId>
-                <accepted>true</accepted>
-                <status>Order Accepted</status>
-            </orderAck>
-        </m1:submitOrderResponse>
-    </SOAP-ENV:Body>
-</SOAP-ENV:Envelope>
