+++ transform-smooks/src/main/resources/smooks/OrderAck_XML.xml
@@ -22,13 +22,13 @@
                       xmlns:jb="http://www.milyn.org/xsd/smooks/javabean-1.4.xsd">
 
     <core:namespaces>
-        <core:namespace prefix="ord" uri="urn:switchyard-quickstarts:transform-smooks:1.0"/>
+        <core:namespace prefix="orders" uri="urn:switchyard-quickstart:transform-smooks:1.0"/>
     </core:namespaces>
 
-    <jb:bean beanId="orderAck" class="org.switchyard.quickstarts.transform.smooks.OrderAck" createOnElement="/ord:submitOrderResponse">
-        <jb:value property="orderId" data="/ord:submitOrderResponse/orderAck/orderId"/>
-        <jb:value property="accepted" data="/ord:submitOrderResponse/orderAck/accepted"/>
-        <jb:value property="status" data="/ord:submitOrderResponse/orderAck/status"/>
+    <jb:bean beanId="orderAck" class="org.switchyard.quickstarts.transform.smooks.OrderAck" createOnElement="/orders:submitOrderResponse">
+        <jb:value property="orderId" data="/orders:submitOrderResponse/orderAck/orderId"/>
+        <jb:value property="accepted" data="/orders:submitOrderResponse/orderAck/accepted"/>
+        <jb:value property="status" data="/orders:submitOrderResponse/orderAck/status"/>
     </jb:bean>
 
 </smooks-resource-list>
